Refactor game into modules
